requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.9
kombu==4.6.10
Markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
python3-openid==3.1.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.21
whitenoise==5.1.0